rename "api" module to "interfaces" use logger better manage.py
