+++ manage.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # coding=utf-8
 
-#  Copyright (c) 2001-2014, Canal TP and/or its affiliates. All rights reserved.
+# Copyright (c) 2001-2016, Canal TP and/or its affiliates. All rights reserved.
 #
 # This file is part of Navitia,
 #     the software to build cool stuff with public transport.
@@ -31,13 +31,13 @@
 # www.navitia.io
 
 from tartare import app
-from tartare.api.grid_calendar import GridCalendar
 import sys
-from flask_restful import Api
+from flask_script import Manager
+from tartare.api import api
 
-api = Api(app)
-api.add_resource(GridCalendar, '/grid_calendar')
+
+manager = Manager(app)
 
 if __name__ == '__main__':
     app.config['DEBUG'] = True
-    app.run()
+    manager.run()
