+++ tartare/interfaces/grid_calendar.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+#coding: utf-8
+
 # Copyright (c) 2001-2016, Canal TP and/or its affiliates. All rights reserved.
 #
 # This file is part of Navitia,
@@ -28,35 +31,12 @@
 # www.navitia.io
 
 import logging
-import logging.config
-import celery
-import sys
+from flask.globals import request
+from flask_restful import Resource
 
-def configure_logger(app_config):
-    """
-    initialize logging
-    """
-    if 'LOGGER' in app_config:
-        logging.config.dictConfig(app_config['LOGGER'])
-    else:  # Default is std out
-        logging.basicConfig(level='INFO')
-
-def make_celery(app):
-    celery_app = celery.Celery(app.import_name,
-                               broker=app.config['CELERY_BROKER_URL'])
-    celery_app.conf.update(app.config)
-    TaskBase = celery_app.Task
-
-    class ContextTask(TaskBase):
-        abstract = True
-
-        def __init__(self):
-            pass
-
-        def __call__(self, *args, **kwargs):
-            with app.app_context():
-                return TaskBase.__call__(self, *args, **kwargs)
-
-    celery_app.Task = ContextTask
-    return celery_app
-
+class GridCalendar(Resource):
+    def post(self):
+        content = request.data
+        logger = logging.getLogger(__name__)
+        logger.info('content received: {}'.format(content))
+        return {'status': 'OK'}, 200
