+++ tartare/tasks.py
@@ -14,7 +14,7 @@ def update_data_task():
     output_dir = app.config.get("OUTPUT_DIR")
     current_data_dir = app.config.get("CURRENT_DATA_DIR")
     logger = logging.getLogger(__name__)
-    logging.info('scanning directory %s', input_dir)
+    logger.info('scanning directory %s', input_dir)
     handle_data(input_dir, output_dir, current_data_dir)
 
 def type_of_data(filename):
